Move PackageSetIterator impl into .cpp
