+++ libdnf/rpm/package_set_iterator.cpp
@@ -18,10 +18,30 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 */
 
 
-#include "libdnf/rpm/package_set_iterator_impl.hpp"
+#include "libdnf/rpm/package_set_impl.hpp"
+#include "libdnf/rpm/package_set_iterator.hpp"
+#include "libdnf/solv/solv_map.hpp"
 
 
 namespace libdnf::rpm {
+
+class PackageSetIterator::Impl : public libdnf::solv::SolvMap::iterator {
+public:
+    Impl(const PackageSet & package_set)
+    : libdnf::solv::SolvMap::iterator(package_set.p_impl->get_map()),
+      package_set{package_set} {}
+
+    Impl(const PackageSetIterator::Impl & package_set_iterator_impl) = default;
+
+    PackageSetIterator::Impl & operator++() {
+        libdnf::solv::SolvMap::iterator::operator++();
+        return *this;
+    }
+
+private:
+    friend PackageSetIterator;
+    const PackageSet & package_set;
+};
 
 
 PackageSetIterator::PackageSetIterator(const PackageSet & package_set) : p_impl{new Impl(package_set)} {}
